Console but it doesnt work

+++ Blastia/Main/GameState/NoisePass.cs
@@ -12,6 +12,6 @@ public class NoisePass
     // strength/intensity
     public float Amplitude { get; set; }
     public float HeightScale { get; set; }
-    public int MaxHeight { get; set; }
+    public float MaxHeight { get; set; }
     public ushort Block { get; set; }
 }
